+++ pyproject.toml
@@ -36,11 +36,11 @@ dependencies = [
     "scikit-learn (>=1.7.0,<2.0.0)",
     "rich (>=14.0.0,<15.0.0)",
     "optuna (>=4.4.0,<5.0.0)",
-    "mlflow (>=3.1.4,<4.0.0)",
     "optuna-integration (>=4.4.0,<5.0.0)",
     "pydantic (>=2.11.7,<3.0.0)",
     "litellm (>=1.75.8,<2.0.0)",
-    "openai (==1.99.9)"
+    "openai (==1.99.9)",
+    "optuna-dashboard (>=0.19.0,<0.20.0)"
 ]
 
 [tool.black]
